long

+++ app.py
@@ -125,7 +125,7 @@ def addOne():
 def subOne():
 	nfc = request.form["nfc"]
 	userId = request.form["userId"]
-	foodRecord = db.food.find_one({"nfc":long(nfc)})
+	foodRecord = db.food.find_one({"nfc":nfc})
 	name = foodRecord["name"]
 	upc = foodRecord["upc"]
 	category = foodRecord["category"]
